refactor: consolidate OpenAI Agents context capture
