+++ aws-opentelemetry-distro/src/amazon/opentelemetry/distro/instrumentation/openai_agents/__init__.py
@@ -16,7 +16,7 @@ from opentelemetry.instrumentation.utils import suppress_http_instrumentation
 class OpenAIAgentsInstrumentor(BaseInstrumentor):  # type: ignore
     """Instrument OpenAI Agents SDK tracing callbacks with OpenTelemetry spans."""
 
-    _wrapper = None
+    _context_capture = None
     _processor = None
     _previous_processors = None
 
@@ -35,38 +35,38 @@ class OpenAIAgentsInstrumentor(BaseInstrumentor):  # type: ignore
             set_trace_processors,
         )
 
+        from ._gen_ai_context_capture import GenAIContextCapture  # pylint: disable=import-outside-toplevel
         from ._processor import OpenTelemetryTracingProcessor  # pylint: disable=import-outside-toplevel
-        from ._wrappers import OpenAIAgentWrapper  # pylint: disable=import-outside-toplevel
 
         tracer_provider = kwargs.get("tracer_provider") or trace.get_tracer_provider()
         tracer = trace.get_tracer(__name__, __version__, tracer_provider=tracer_provider)
         self._processor = OpenTelemetryTracingProcessor(tracer)
-        wrapper = self._wrapper = OpenAIAgentWrapper(self._processor)
+        context_capture = self._context_capture = GenAIContextCapture()
 
         try_wrap(
             "openai.resources.responses.responses",
             "Responses.create",
-            wrapper.capture_openai_request_attributes,
+            context_capture.capture_openai_request_attributes,
         )
         try_wrap(
             "openai.resources.responses.responses",
             "AsyncResponses.create",
-            wrapper.capture_openai_request_attributes,
+            context_capture.capture_openai_request_attributes,
         )
         try_wrap(
             "openai.resources.chat.completions.completions",
             "Completions.create",
-            wrapper.capture_openai_completion_attributes,
+            context_capture.capture_openai_completion_attributes,
         )
         try_wrap(
             "openai.resources.chat.completions.completions",
             "AsyncCompletions.create",
-            wrapper.capture_openai_completion_attributes,
+            context_capture.capture_openai_completion_attributes,
         )
-        try_wrap("litellm", "completion", wrapper.capture_litellm_completion_attributes)
-        try_wrap("litellm", "acompletion", wrapper.capture_litellm_completion_attributes)
-        try_wrap("agents.items", "ItemHelpers.tool_call_output_item", wrapper.capture_tool_call_attributes)
-        try_wrap("agents.tool_context", "ToolContext.from_agent_context", wrapper.capture_tool_call_attributes)
+        try_wrap("litellm", "completion", context_capture.capture_litellm_completion_attributes)
+        try_wrap("litellm", "acompletion", context_capture.capture_litellm_completion_attributes)
+        try_wrap("agents.items", "ItemHelpers.tool_call_output_item", context_capture.capture_tool_call_attributes)
+        try_wrap("agents.tool_context", "ToolContext.from_agent_context", context_capture.capture_tool_call_attributes)
         # disables http spans created from spans sent OpenAI's tracing backend
         try_wrap("agents.tracing.processors", "BackendSpanExporter.export", _suppress_http_instrumentation)
 
@@ -107,7 +107,7 @@ class OpenAIAgentsInstrumentor(BaseInstrumentor):  # type: ignore
                 set_trace_processors([item for item in current_processors if item is not processor])
             processor.shutdown()
         finally:
-            self._wrapper = None
+            self._context_capture = None
             self._processor = None
             self._previous_processors = None
