+++ aws-opentelemetry-distro/src/amazon/opentelemetry/distro/instrumentation/openai_agents/_processor.py
@@ -9,6 +9,7 @@ from collections.abc import Mapping, Sequence
 from contextvars import Token
 from dataclasses import dataclass
 from typing import Any, Optional, Union
+from urllib.parse import urlparse
 
 from agents.tracing import Span as AgentsSpan
 from agents.tracing import Trace as AgentsTrace
@@ -24,6 +25,7 @@ from pydantic import BaseModel
 from typing_extensions import override
 
 from amazon.opentelemetry.distro.instrumentation.common.instrumentation_utils import (
+    PROVIDER_MAP,
     DictWithLock,
     content_to_parts,
     first_not_none,
@@ -32,34 +34,58 @@ from amazon.opentelemetry.distro.instrumentation.common.instrumentation_utils im
     to_tool_attribute_value,
     try_detach,
 )
-from amazon.opentelemetry.distro.instrumentation.openai_agents._telemetry_helpers import _TelemetryHelpers
 from opentelemetry import context
 from opentelemetry.semconv._incubating.attributes.gen_ai_attributes import (
     GEN_AI_AGENT_NAME,
     GEN_AI_INPUT_MESSAGES,
+    GEN_AI_OPENAI_REQUEST_SERVICE_TIER,
+    GEN_AI_OPENAI_RESPONSE_SERVICE_TIER,
+    GEN_AI_OPENAI_RESPONSE_SYSTEM_FINGERPRINT,
     GEN_AI_OPERATION_NAME,
     GEN_AI_OUTPUT_MESSAGES,
     GEN_AI_OUTPUT_TYPE,
     GEN_AI_PROVIDER_NAME,
+    GEN_AI_REQUEST_CHOICE_COUNT,
+    GEN_AI_REQUEST_FREQUENCY_PENALTY,
+    GEN_AI_REQUEST_MAX_TOKENS,
     GEN_AI_REQUEST_MODEL,
+    GEN_AI_REQUEST_PRESENCE_PENALTY,
+    GEN_AI_REQUEST_SEED,
+    GEN_AI_REQUEST_STOP_SEQUENCES,
     GEN_AI_REQUEST_STREAM,
+    GEN_AI_REQUEST_TEMPERATURE,
+    GEN_AI_REQUEST_TOP_K,
+    GEN_AI_REQUEST_TOP_P,
     GEN_AI_RESPONSE_FINISH_REASONS,
+    GEN_AI_RESPONSE_ID,
     GEN_AI_RESPONSE_MODEL,
     GEN_AI_SYSTEM_INSTRUCTIONS,
     GEN_AI_TOOL_CALL_ARGUMENTS,
+    GEN_AI_TOOL_CALL_ID,
     GEN_AI_TOOL_CALL_RESULT,
+    GEN_AI_TOOL_DEFINITIONS,
     GEN_AI_TOOL_NAME,
     GEN_AI_TOOL_TYPE,
+    GEN_AI_USAGE_CACHE_CREATION_INPUT_TOKENS,
+    GEN_AI_USAGE_CACHE_READ_INPUT_TOKENS,
+    GEN_AI_USAGE_INPUT_TOKENS,
+    GEN_AI_USAGE_OUTPUT_TOKENS,
+    GEN_AI_USAGE_REASONING_OUTPUT_TOKENS,
     GEN_AI_WORKFLOW_NAME,
     GenAiOperationNameValues,
     GenAiOutputTypeValues,
     GenAiProviderNameValues,
 )
 from opentelemetry.semconv.attributes.error_attributes import ERROR_TYPE
+from opentelemetry.semconv.attributes.server_attributes import SERVER_ADDRESS, SERVER_PORT
 from opentelemetry.trace import Span, SpanKind, Status, StatusCode, Tracer, set_span_in_context
 from opentelemetry.util.types import AttributeValue
 
+from ._gen_ai_context_capture import GenAIContextCapture
+
 _logger = logging.getLogger(__name__)
+
+GEN_AI_REQUEST_REASONING_LEVEL = "gen_ai.request.reasoning.level"
 
 
 @dataclass
@@ -89,10 +115,6 @@ class OpenTelemetryTracingProcessor(TracingProcessor):
         # Maps OpenAI span IDs to OTel spans, context tokens, and shared agent content.
         # Used to resolve parents, finish spans, and carry state through unhandled spans.
         self._openai_span_id_to_otel_span_entry = DictWithLock()
-
-    def get_otel_span(self, openai_span_id: str) -> Optional[Span]:
-        entry = self._openai_span_id_to_otel_span_entry.get(openai_span_id)
-        return entry.span if entry is not None else None
 
     @override
     def on_trace_start(self, trace: AgentsTrace) -> None:
@@ -133,6 +155,12 @@ class OpenTelemetryTracingProcessor(TracingProcessor):
                 )
             return
 
+        if isinstance(span_data, (GenerationSpanData, ResponseSpanData)):
+            model_config = get_value(span_data, "model_config") or {}
+            GenAIContextCapture.reset_current_model_context(get_value(model_config, "model_impl"))
+        if isinstance(span_data, (FunctionSpanData, HandoffSpanData)):
+            GenAIContextCapture.reset_current_tool_context()
+
         agent_content = parent_entry.agent_content if parent_entry is not None else None
         if isinstance(span_data, AgentSpanData):
             agent_content = _AgentContent()
@@ -150,19 +178,21 @@ class OpenTelemetryTracingProcessor(TracingProcessor):
                     value = extra_args.get(key)
                     if value is not None and not model_config.get(key):
                         model_config[key] = value
-            _TelemetryHelpers.set_attribute(attributes, GEN_AI_REQUEST_MODEL, span_data.model)
-            _TelemetryHelpers.set_attribute(
+            OpenTelemetryTracingProcessor._set_attribute(attributes, GEN_AI_REQUEST_MODEL, span_data.model)
+            OpenTelemetryTracingProcessor._set_attribute(
                 attributes,
                 GEN_AI_PROVIDER_NAME,
-                _TelemetryHelpers.resolve_provider(
+                OpenTelemetryTracingProcessor._resolve_provider(
                     span_data.model,
                     model_config.get("base_url") or model_config.get("api_base"),
                     model_config.get("custom_llm_provider"),
                 ),
             )
-            _TelemetryHelpers.set_request_attributes(attributes, model_config)
+            OpenTelemetryTracingProcessor._set_request_attributes(attributes, model_config)
         elif isinstance(span_data, ResponseSpanData):
-            _TelemetryHelpers.set_attribute(attributes, GEN_AI_REQUEST_MODEL, get_value(span_data.response, "model"))
+            OpenTelemetryTracingProcessor._set_attribute(
+                attributes, GEN_AI_REQUEST_MODEL, get_value(span_data.response, "model")
+            )
         otel_span = self._tracer.start_span(
             self._set_span_name(span_data, operation),
             context=set_span_in_context(parent_entry.span) if parent_entry is not None else None,
@@ -190,15 +220,7 @@ class OpenTelemetryTracingProcessor(TracingProcessor):
         otel_span = entry.span
 
         try:
-            current_span_attributes = dict(get_value(otel_span, "attributes") or {})
-            processor_attributes, content = self._set_span_attributes(
-                span,
-                entry.agent_content,
-                current_span_attributes,
-            )
-            # Preserve attributes already set on the span; processor-derived values are fallbacks.
-            attributes = dict(processor_attributes)
-            attributes.update(current_span_attributes)
+            attributes, content = self._set_span_attributes(span, entry.agent_content)
             if isinstance(span_data, (GenerationSpanData, ResponseSpanData)):
                 operation = self._set_operation_name(span_data)
                 attributes[GEN_AI_OPERATION_NAME] = operation
@@ -309,13 +331,12 @@ class OpenTelemetryTracingProcessor(TracingProcessor):
         self,
         span: AgentsSpan[Any],
         agent_content: Optional[_AgentContent],
-        current_span_attributes: Mapping[str, AttributeValue],
     ) -> tuple[dict[str, AttributeValue], Optional[_AgentContent]]:
         span_data = span.span_data
         if isinstance(span_data, GenerationSpanData):
-            return self._set_generation_attributes(span_data, current_span_attributes)
+            return self._set_generation_attributes(span_data)
         if isinstance(span_data, ResponseSpanData):
-            return self._set_response_attributes(span_data)
+            return self._set_response_span_attributes(span_data)
         if isinstance(span_data, FunctionSpanData):
             return self._set_function_attributes(span_data), None
         if isinstance(span_data, HandoffSpanData):
@@ -327,7 +348,6 @@ class OpenTelemetryTracingProcessor(TracingProcessor):
     @staticmethod
     def _set_generation_attributes(
         span_data: GenerationSpanData,
-        current_span_attributes: Mapping[str, AttributeValue],
     ) -> tuple[dict[str, AttributeValue], _AgentContent]:
         attributes: dict[str, AttributeValue] = {}
         streamed_response = OpenTelemetryTracingProcessor._get_response_payload(span_data.output)
@@ -336,23 +356,10 @@ class OpenTelemetryTracingProcessor(TracingProcessor):
             **(span_data.model_config or {}),
         }
 
-        request_model = span_data.model
-        _TelemetryHelpers.set_attribute(attributes, GEN_AI_REQUEST_MODEL, request_model)
-        _TelemetryHelpers.set_attribute(
-            attributes,
-            GEN_AI_PROVIDER_NAME,
-            _TelemetryHelpers.resolve_provider(
-                request_model,
-                model_config.get("base_url") or model_config.get("api_base"),
-                model_config.get("custom_llm_provider"),
-            ),
-        )
-        _TelemetryHelpers.set_request_attributes(attributes, model_config)
-
         output_items = span_data.output
         if streamed_response is not None:
-            _TelemetryHelpers.set_response_attributes(attributes, streamed_response)
-            _TelemetryHelpers.set_attribute(attributes, GEN_AI_REQUEST_STREAM, True)
+            OpenTelemetryTracingProcessor._set_response_payload_attributes(attributes, streamed_response)
+            OpenTelemetryTracingProcessor._set_attribute(attributes, GEN_AI_REQUEST_STREAM, True)
             output_items = get_value(streamed_response, "output")
 
         span_usage = span_data.usage
@@ -361,15 +368,19 @@ class OpenTelemetryTracingProcessor(TracingProcessor):
             for key in ("input_tokens", "prompt_tokens", "output_tokens", "completion_tokens")
         ):
             span_usage = None
-        _TelemetryHelpers.set_usage_attributes(
+        OpenTelemetryTracingProcessor._set_usage_attributes(
             attributes,
             first_not_none(span_usage, get_value(streamed_response, "usage")),
             first_not_none(get_value(streamed_response, "usage"), span_usage),
         )
-        current_finish_reasons = list(current_span_attributes.get(GEN_AI_RESPONSE_FINISH_REASONS, ()))
+        captured_finish_reasons = OpenTelemetryTracingProcessor._set_current_model_context_attributes(
+            attributes,
+            model_config,
+            span_data.model,
+        )
         default_finish_reason = first_not_none(
-            current_finish_reasons[0] if current_finish_reasons else None,
-            _TelemetryHelpers.get_finish_reason(streamed_response, attributes),
+            captured_finish_reasons[0] if captured_finish_reasons else None,
+            OpenTelemetryTracingProcessor._get_finish_reason(streamed_response, attributes),
         )
 
         system_instructions, input_messages = _GenAIMessageNormalizer.normalize_input_messages(span_data.input)
@@ -377,9 +388,11 @@ class OpenTelemetryTracingProcessor(TracingProcessor):
         OpenTelemetryTracingProcessor._set_message_attributes(
             attributes, system_instructions, input_messages, output_messages
         )
-        if GEN_AI_RESPONSE_FINISH_REASONS not in current_span_attributes:
+        if not captured_finish_reasons:
             finish_reasons = [message["finish_reason"] for message in output_messages if message.get("finish_reason")]
-            _TelemetryHelpers.set_attribute(attributes, GEN_AI_RESPONSE_FINISH_REASONS, finish_reasons or None)
+            OpenTelemetryTracingProcessor._set_attribute(
+                attributes, GEN_AI_RESPONSE_FINISH_REASONS, finish_reasons or None
+            )
 
         content = _AgentContent(
             input_messages=input_messages or None,
@@ -390,30 +403,35 @@ class OpenTelemetryTracingProcessor(TracingProcessor):
         return attributes, content
 
     @staticmethod
-    def _set_response_attributes(span_data: ResponseSpanData) -> tuple[dict[str, AttributeValue], _AgentContent]:
+    def _set_response_span_attributes(span_data: ResponseSpanData) -> tuple[dict[str, AttributeValue], _AgentContent]:
         attributes: dict[str, AttributeValue] = {}
         response = span_data.response
         model_config = OpenTelemetryTracingProcessor._get_response_request_config(response)
 
         response_model = get_value(response, "model")
         request_model = response_model
-        _TelemetryHelpers.set_response_attributes(attributes, response)
-        _TelemetryHelpers.set_attribute(attributes, GEN_AI_REQUEST_MODEL, request_model)
-        _TelemetryHelpers.set_attribute(
+        OpenTelemetryTracingProcessor._set_response_payload_attributes(attributes, response)
+        OpenTelemetryTracingProcessor._set_attribute(attributes, GEN_AI_REQUEST_MODEL, request_model)
+        OpenTelemetryTracingProcessor._set_attribute(
             attributes,
             GEN_AI_PROVIDER_NAME,
-            _TelemetryHelpers.resolve_provider(
+            OpenTelemetryTracingProcessor._resolve_provider(
                 request_model,
                 model_config.get("base_url") or model_config.get("api_base"),
                 model_config.get("custom_llm_provider"),
             ),
         )
-        _TelemetryHelpers.set_request_attributes(attributes, model_config)
+        OpenTelemetryTracingProcessor._set_request_attributes(attributes, model_config)
 
-        _TelemetryHelpers.set_usage_attributes(
+        OpenTelemetryTracingProcessor._set_usage_attributes(
             attributes,
             first_not_none(getattr(span_data, "usage", None), get_value(response, "usage")),
             first_not_none(get_value(response, "usage"), getattr(span_data, "usage", None)),
+        )
+        captured_finish_reasons = OpenTelemetryTracingProcessor._set_current_model_context_attributes(
+            attributes,
+            model_config,
+            response_model,
         )
 
         system_instructions, input_messages = _GenAIMessageNormalizer.normalize_input_messages(
@@ -424,15 +442,21 @@ class OpenTelemetryTracingProcessor(TracingProcessor):
         )
         if response_instructions:
             system_instructions = response_instructions
-        default_finish_reason = _TelemetryHelpers.get_finish_reason(response, attributes)
+        default_finish_reason = first_not_none(
+            captured_finish_reasons[0] if captured_finish_reasons else None,
+            OpenTelemetryTracingProcessor._get_finish_reason(response, attributes),
+        )
         output_messages = _GenAIMessageNormalizer.normalize_output_messages(
             get_value(response, "output"), default_finish_reason
         )
         OpenTelemetryTracingProcessor._set_message_attributes(
             attributes, system_instructions, input_messages, output_messages
         )
-        finish_reasons = [message["finish_reason"] for message in output_messages if message.get("finish_reason")]
-        _TelemetryHelpers.set_attribute(attributes, GEN_AI_RESPONSE_FINISH_REASONS, finish_reasons or None)
+        if not captured_finish_reasons:
+            finish_reasons = [message["finish_reason"] for message in output_messages if message.get("finish_reason")]
+            OpenTelemetryTracingProcessor._set_attribute(
+                attributes, GEN_AI_RESPONSE_FINISH_REASONS, finish_reasons or None
+            )
 
         content = _AgentContent(
             input_messages=input_messages or None,
@@ -445,22 +469,34 @@ class OpenTelemetryTracingProcessor(TracingProcessor):
     @staticmethod
     def _set_function_attributes(span_data: FunctionSpanData) -> dict[str, AttributeValue]:
         attributes: dict[str, AttributeValue] = {}
-        _TelemetryHelpers.set_attribute(attributes, GEN_AI_TOOL_NAME, span_data.name)
-        _TelemetryHelpers.set_attribute(attributes, GEN_AI_TOOL_TYPE, "function")
-        _TelemetryHelpers.set_attribute(
+        OpenTelemetryTracingProcessor._set_attribute(attributes, GEN_AI_TOOL_NAME, span_data.name)
+        OpenTelemetryTracingProcessor._set_attribute(attributes, GEN_AI_TOOL_TYPE, "function")
+        OpenTelemetryTracingProcessor._set_attribute(
+            attributes,
+            GEN_AI_TOOL_CALL_ID,
+            get_value(GenAIContextCapture.get_current_tool_context(), "call_id"),
+        )
+        OpenTelemetryTracingProcessor._set_attribute(
             attributes, GEN_AI_TOOL_CALL_ARGUMENTS, to_tool_attribute_value(span_data.input)
         )
-        _TelemetryHelpers.set_attribute(attributes, GEN_AI_TOOL_CALL_RESULT, to_tool_attribute_value(span_data.output))
+        OpenTelemetryTracingProcessor._set_attribute(
+            attributes, GEN_AI_TOOL_CALL_RESULT, to_tool_attribute_value(span_data.output)
+        )
         return attributes
 
     @staticmethod
     def _set_handoff_attributes(span_data: HandoffSpanData) -> dict[str, AttributeValue]:
         attributes: dict[str, AttributeValue] = {}
-        _TelemetryHelpers.set_attribute(
+        OpenTelemetryTracingProcessor._set_attribute(
             attributes, GEN_AI_TOOL_NAME, OpenTelemetryTracingProcessor._get_handoff_tool_name(span_data)
         )
-        _TelemetryHelpers.set_attribute(attributes, GEN_AI_TOOL_TYPE, "function")
-        _TelemetryHelpers.set_attribute(
+        OpenTelemetryTracingProcessor._set_attribute(attributes, GEN_AI_TOOL_TYPE, "function")
+        OpenTelemetryTracingProcessor._set_attribute(
+            attributes,
+            GEN_AI_TOOL_CALL_ID,
+            get_value(GenAIContextCapture.get_current_tool_context(), "call_id"),
+        )
+        OpenTelemetryTracingProcessor._set_attribute(
             attributes,
             GEN_AI_TOOL_CALL_ARGUMENTS,
             to_tool_attribute_value({"from_agent": span_data.from_agent, "to_agent": span_data.to_agent}),
@@ -473,12 +509,12 @@ class OpenTelemetryTracingProcessor(TracingProcessor):
         span_data: AgentSpanData,
     ) -> dict[str, AttributeValue]:
         attributes: dict[str, AttributeValue] = {}
-        _TelemetryHelpers.set_attribute(attributes, GEN_AI_AGENT_NAME, span_data.name)
+        OpenTelemetryTracingProcessor._set_attribute(attributes, GEN_AI_AGENT_NAME, span_data.name)
         content = content or _AgentContent()
         for key, value in (content.request_attributes or {}).items():
-            _TelemetryHelpers.set_attribute(attributes, key, value)
-        _TelemetryHelpers.set_tool_definitions(attributes, span_data.tools)
-        _TelemetryHelpers.set_attribute(
+            OpenTelemetryTracingProcessor._set_attribute(attributes, key, value)
+        OpenTelemetryTracingProcessor._set_tool_definitions(attributes, span_data.tools)
+        OpenTelemetryTracingProcessor._set_attribute(
             attributes, GEN_AI_OUTPUT_TYPE, self._get_agent_output_type(span_data.output_type)
         )
         self._set_message_attributes(
@@ -488,6 +524,356 @@ class OpenTelemetryTracingProcessor(TracingProcessor):
             content.output_messages or [],
         )
         return attributes
+
+    @staticmethod
+    def _set_current_model_context_attributes(
+        attributes: dict[str, AttributeValue],
+        model_config: Mapping[str, Any],
+        fallback_model: Any,
+    ) -> list[str]:
+        current_model_context = GenAIContextCapture.get_current_model_context()
+        request_params = get_value(current_model_context, "request_params") or {}
+        model_config = {**model_config, **request_params}
+        request_model = first_not_none(request_params.get("model"), fallback_model)
+        OpenTelemetryTracingProcessor._set_attribute(attributes, GEN_AI_REQUEST_MODEL, request_model)
+        OpenTelemetryTracingProcessor._set_attribute(
+            attributes,
+            GEN_AI_PROVIDER_NAME,
+            OpenTelemetryTracingProcessor._resolve_provider(
+                request_model,
+                model_config.get("base_url") or model_config.get("api_base"),
+                model_config.get("custom_llm_provider"),
+            ),
+        )
+        OpenTelemetryTracingProcessor._set_request_attributes(attributes, model_config)
+
+        finish_reasons: list[str] = []
+        for response in get_value(current_model_context, "responses") or []:
+            OpenTelemetryTracingProcessor._set_response_payload_attributes(attributes, response)
+            usage = get_value(response, "usage")
+            OpenTelemetryTracingProcessor._set_usage_attributes(attributes, usage, usage)
+            finish_reasons.extend(OpenTelemetryTracingProcessor._get_finish_reasons(response, attributes))
+        OpenTelemetryTracingProcessor._set_attribute(
+            attributes,
+            GEN_AI_RESPONSE_FINISH_REASONS,
+            finish_reasons or None,
+        )
+        return finish_reasons
+
+    @staticmethod
+    def _set_request_attributes(attributes: dict[str, AttributeValue], model_config: Mapping[str, Any]) -> None:
+        OpenTelemetryTracingProcessor._set_attribute(
+            attributes,
+            GEN_AI_REQUEST_TEMPERATURE,
+            model_config.get("temperature"),
+        )
+        OpenTelemetryTracingProcessor._set_attribute(attributes, GEN_AI_REQUEST_TOP_P, model_config.get("top_p"))
+        OpenTelemetryTracingProcessor._set_attribute(attributes, GEN_AI_REQUEST_TOP_K, model_config.get("top_k"))
+        OpenTelemetryTracingProcessor._set_attribute(
+            attributes,
+            GEN_AI_REQUEST_MAX_TOKENS,
+            first_not_none(
+                model_config.get("max_tokens"),
+                model_config.get("max_output_tokens"),
+                model_config.get("max_completion_tokens"),
+            ),
+        )
+        OpenTelemetryTracingProcessor._set_attribute(
+            attributes,
+            GEN_AI_REQUEST_FREQUENCY_PENALTY,
+            model_config.get("frequency_penalty"),
+        )
+        OpenTelemetryTracingProcessor._set_attribute(
+            attributes,
+            GEN_AI_REQUEST_PRESENCE_PENALTY,
+            model_config.get("presence_penalty"),
+        )
+        OpenTelemetryTracingProcessor._set_attribute(
+            attributes,
+            GEN_AI_REQUEST_STOP_SEQUENCES,
+            OpenTelemetryTracingProcessor._to_string_sequence(
+                first_not_none(model_config.get("stop_sequences"), model_config.get("stop"))
+            ),
+        )
+        OpenTelemetryTracingProcessor._set_attribute(attributes, GEN_AI_REQUEST_SEED, model_config.get("seed"))
+        OpenTelemetryTracingProcessor._set_attribute(
+            attributes,
+            GEN_AI_REQUEST_CHOICE_COUNT,
+            first_not_none(model_config.get("choice_count"), model_config.get("n")),
+        )
+        OpenTelemetryTracingProcessor._set_attribute(attributes, GEN_AI_REQUEST_STREAM, model_config.get("stream"))
+        OpenTelemetryTracingProcessor._set_attribute(
+            attributes,
+            GEN_AI_OPENAI_REQUEST_SERVICE_TIER,
+            model_config.get("service_tier"),
+        )
+
+        base_url = model_config.get("base_url") or model_config.get("api_base")
+        if base_url:
+            parsed_url = urlparse(str(base_url))
+            OpenTelemetryTracingProcessor._set_attribute(attributes, SERVER_ADDRESS, parsed_url.hostname)
+            OpenTelemetryTracingProcessor._set_attribute(attributes, SERVER_PORT, parsed_url.port)
+
+        OpenTelemetryTracingProcessor._set_tool_definitions(
+            attributes,
+            model_config.get("tools") or model_config.get("functions"),
+        )
+        OpenTelemetryTracingProcessor._set_attribute(
+            attributes, GEN_AI_OUTPUT_TYPE, OpenTelemetryTracingProcessor._get_request_output_type(model_config)
+        )
+        OpenTelemetryTracingProcessor._set_attribute(
+            attributes,
+            GEN_AI_REQUEST_REASONING_LEVEL,
+            first_not_none(model_config.get("reasoning_effort"), get_value(model_config.get("reasoning"), "effort")),
+        )
+
+    @staticmethod
+    def _set_response_payload_attributes(attributes: dict[str, AttributeValue], response: Any) -> None:
+        response_id = get_value(response, "id")
+        OpenTelemetryTracingProcessor._set_attribute(
+            attributes, GEN_AI_RESPONSE_ID, None if response_id in ("__fake_id__", "") else response_id
+        )
+        OpenTelemetryTracingProcessor._set_attribute(attributes, GEN_AI_RESPONSE_MODEL, get_value(response, "model"))
+        OpenTelemetryTracingProcessor._set_attribute(
+            attributes,
+            GEN_AI_OPENAI_RESPONSE_SERVICE_TIER,
+            get_value(response, "service_tier"),
+        )
+        OpenTelemetryTracingProcessor._set_attribute(
+            attributes,
+            GEN_AI_OPENAI_RESPONSE_SYSTEM_FINGERPRINT,
+            get_value(response, "system_fingerprint"),
+        )
+
+    @staticmethod
+    def _set_usage_attributes(attributes: dict[str, AttributeValue], usage: Any, detailed_usage: Any = None) -> None:
+        detailed_usage = first_not_none(detailed_usage, usage)
+        input_details = first_not_none(
+            get_value(detailed_usage, "input_tokens_details"),
+            get_value(detailed_usage, "prompt_tokens_details"),
+            get_value(usage, "input_tokens_details"),
+            get_value(usage, "prompt_tokens_details"),
+        )
+        output_details = first_not_none(
+            get_value(detailed_usage, "output_tokens_details"),
+            get_value(detailed_usage, "completion_tokens_details"),
+            get_value(usage, "output_tokens_details"),
+            get_value(usage, "completion_tokens_details"),
+        )
+        OpenTelemetryTracingProcessor._set_attribute(
+            attributes,
+            GEN_AI_USAGE_INPUT_TOKENS,
+            first_not_none(
+                get_value(usage, "input_tokens"),
+                get_value(usage, "prompt_tokens"),
+                get_value(detailed_usage, "input_tokens"),
+                get_value(detailed_usage, "prompt_tokens"),
+            ),
+        )
+        OpenTelemetryTracingProcessor._set_attribute(
+            attributes,
+            GEN_AI_USAGE_OUTPUT_TOKENS,
+            first_not_none(
+                get_value(usage, "output_tokens"),
+                get_value(usage, "completion_tokens"),
+                get_value(detailed_usage, "output_tokens"),
+                get_value(detailed_usage, "completion_tokens"),
+            ),
+        )
+        OpenTelemetryTracingProcessor._set_attribute(
+            attributes,
+            GEN_AI_USAGE_CACHE_READ_INPUT_TOKENS,
+            first_not_none(
+                get_value(input_details, "cached_tokens"),
+                get_value(detailed_usage, "cache_read_input_tokens"),
+                get_value(usage, "cache_read_input_tokens"),
+            ),
+        )
+        OpenTelemetryTracingProcessor._set_attribute(
+            attributes,
+            GEN_AI_USAGE_CACHE_CREATION_INPUT_TOKENS,
+            first_not_none(
+                get_value(input_details, "cache_write_tokens"),
+                get_value(detailed_usage, "cache_write_input_tokens"),
+                get_value(detailed_usage, "cache_creation_input_tokens"),
+                get_value(usage, "cache_write_input_tokens"),
+                get_value(usage, "cache_creation_input_tokens"),
+            ),
+        )
+        OpenTelemetryTracingProcessor._set_attribute(
+            attributes,
+            GEN_AI_USAGE_REASONING_OUTPUT_TOKENS,
+            first_not_none(
+                get_value(output_details, "reasoning_tokens"),
+                get_value(detailed_usage, "reasoning_tokens"),
+                get_value(usage, "reasoning_tokens"),
+            ),
+        )
+
+    @staticmethod
+    def _set_tool_definitions(attributes: dict[str, AttributeValue], tools: Any) -> None:
+        definitions: list[dict[str, Any]] = []
+        for tool in OpenTelemetryTracingProcessor._to_item_list(tools):
+            if isinstance(tool, str):
+                definitions.append({"type": "function", "name": tool})
+                continue
+            tool_data = OpenTelemetryTracingProcessor._to_mapping(tool)
+            if tool_data is None:
+                continue
+            function = OpenTelemetryTracingProcessor._to_mapping(tool_data.get("function")) or {}
+            definition: dict[str, Any] = {"type": tool_data.get("type") or "function"}
+            for key, source_key in (("name", "name"), ("description", "description"), ("parameters", "parameters")):
+                value = first_not_none(tool_data.get(source_key), function.get(source_key))
+                if value is not None:
+                    definition[key] = value
+            definitions.append(definition)
+        if definitions:
+            attributes[GEN_AI_TOOL_DEFINITIONS] = serialize_to_json_string(definitions)
+
+    @staticmethod
+    def _get_request_output_type(model_config: Mapping[str, Any]) -> Optional[str]:
+        modalities = OpenTelemetryTracingProcessor._to_string_sequence(model_config.get("modalities")) or []
+        if "audio" in modalities:
+            return GenAiOutputTypeValues.SPEECH.value
+        if "image" in modalities:
+            return GenAiOutputTypeValues.IMAGE.value
+        response_format = first_not_none(
+            get_value(model_config.get("text"), "format"),
+            model_config.get("response_format"),
+        )
+        format_type = first_not_none(get_value(response_format, "type"), response_format)
+        if not isinstance(format_type, str):
+            return None
+        return {
+            "text": GenAiOutputTypeValues.TEXT.value,
+            "json": GenAiOutputTypeValues.JSON.value,
+            "json_object": GenAiOutputTypeValues.JSON.value,
+            "json_schema": GenAiOutputTypeValues.JSON.value,
+        }.get(format_type)
+
+    @staticmethod
+    def _get_finish_reasons(response: Any, attributes: Mapping[str, AttributeValue]) -> list[str]:
+        reasons = OpenTelemetryTracingProcessor._to_item_list(get_value(response, "finish_reasons"))
+        reasons.extend(
+            get_value(choice, "finish_reason")
+            for choice in OpenTelemetryTracingProcessor._to_item_list(get_value(response, "choices"))
+        )
+        finish_reason = get_value(response, "finish_reason")
+        if finish_reason is not None:
+            reasons.append(finish_reason)
+        normalized = [
+            OpenTelemetryTracingProcessor._normalize_finish_reason({"finish_reason": reason}, False)
+            for reason in reasons
+            if reason is not None
+        ]
+        if not normalized:
+            fallback = OpenTelemetryTracingProcessor._get_finish_reason(response, attributes)
+            if fallback is not None:
+                normalized.append(fallback)
+        return normalized
+
+    @staticmethod
+    def _get_finish_reason(response: Any, attributes: Mapping[str, AttributeValue]) -> Optional[str]:
+        reasons = OpenTelemetryTracingProcessor._to_item_list(get_value(response, "finish_reasons"))
+        if not reasons:
+            reasons = [
+                get_value(choice, "finish_reason")
+                for choice in OpenTelemetryTracingProcessor._to_item_list(get_value(response, "choices"))
+            ]
+        for reason in reasons:
+            if reason is not None:
+                return OpenTelemetryTracingProcessor._normalize_finish_reason({"finish_reason": reason}, False)
+
+        reason = first_not_none(
+            get_value(response, "incomplete_reason"),
+            get_value(get_value(response, "incomplete_details"), "reason"),
+        )
+        if reason:
+            return OpenTelemetryTracingProcessor._normalize_finish_reason({"finish_reason": reason}, False)
+        if get_value(response, "status") in ("failed", "cancelled"):
+            return "error"
+        max_tokens = attributes.get(GEN_AI_REQUEST_MAX_TOKENS)
+        output_tokens = attributes.get(GEN_AI_USAGE_OUTPUT_TOKENS)
+        if isinstance(max_tokens, int) and isinstance(output_tokens, int) and output_tokens >= max_tokens > 0:
+            return "length"
+        return None
+
+    @staticmethod
+    def _normalize_finish_reason(
+        item: Mapping[str, Any],
+        has_tool_calls: bool,
+        default_finish_reason: Optional[str] = None,
+    ) -> str:
+        if has_tool_calls:
+            return "tool_call"
+        raw_reason = item.get("finish_reason") or item.get("stop_reason")
+        if raw_reason is None and item.get("status") == "incomplete":
+            return "length"
+        if raw_reason is None:
+            return default_finish_reason or "stop"
+        return {
+            "end_turn": "stop",
+            "stop": "stop",
+            "tool_calls": "tool_call",
+            "tool_use": "tool_call",
+            "max_tokens": "length",
+            "max_output_tokens": "length",
+            "length": "length",
+            "content_filter": "content_filter",
+        }.get(str(raw_reason), str(raw_reason))
+
+    @staticmethod
+    def _resolve_provider(model: Any, base_url: Any, custom_provider: Any = None) -> str:
+        """Resolve known OTel providers while preserving explicit custom provider names."""
+        if custom_provider:
+            candidate = re.sub(r"[^a-z0-9._-]+", "_", str(custom_provider).strip().lower())
+            return PROVIDER_MAP.get(candidate, candidate)
+        if model and "/" in str(model):
+            candidate = re.sub(r"[^a-z0-9._-]+", "_", str(model).split("/", 1)[0].strip().lower())
+            return PROVIDER_MAP.get(candidate, candidate)
+        if base_url:
+            hostname = urlparse(str(base_url)).hostname or ""
+            if hostname.endswith((".services.ai.azure.com", ".models.ai.azure.com")):
+                return GenAiProviderNameValues.AZURE_AI_INFERENCE.value
+            for label in reversed(hostname.split(".")):
+                for candidate in label.split("-"):
+                    provider = PROVIDER_MAP.get(candidate.lower())
+                    if provider is not None:
+                        return provider
+        return GenAiProviderNameValues.OPENAI.value
+
+    @staticmethod
+    def _to_string_sequence(value: Any) -> Optional[list[str]]:
+        if value is None:
+            return None
+        if isinstance(value, (str, bytes)):
+            return [str(value)]
+        if isinstance(value, Sequence):
+            return [str(item) for item in value]
+        return [str(value)]
+
+    @staticmethod
+    def _to_item_list(items: Any) -> list[Any]:
+        if items is None:
+            return []
+        if isinstance(items, (str, bytes, Mapping, BaseModel)):
+            return [items]
+        if isinstance(items, Sequence):
+            return list(items)
+        return [items]
+
+    @staticmethod
+    def _to_mapping(value: Any) -> Optional[dict[str, Any]]:
+        if isinstance(value, Mapping):
+            return dict(value)
+        if isinstance(value, BaseModel):
+            return value.model_dump()
+        return None
+
+    @staticmethod
+    def _set_attribute(attributes: dict[str, AttributeValue], key: str, value: Any) -> None:
+        if isinstance(value, (bool, bytes, float, int, str, list, tuple)):
+            attributes[key] = value
 
     @staticmethod
     def _set_message_attributes(
@@ -564,6 +950,9 @@ class OpenTelemetryTracingProcessor(TracingProcessor):
 
     @staticmethod
     def _get_handoff_tool_name(span_data: HandoffSpanData) -> str:
+        tool_name = get_value(GenAIContextCapture.get_current_tool_context(), "name")
+        if tool_name:
+            return str(tool_name)
         if not span_data.to_agent:
             return "handoff"
         return re.sub(r"[^a-zA-Z0-9_]", "_", f"transfer_to_{span_data.to_agent}").lower()
